requirements.txt
Pillow~=9.1.1
requests~=2.26.0
Flask~=2.1.2